pyproject.toml
[project]
name = "personal_site"
classifiers = ["Private :: Do Not Upload"]
version = "0.1.0"
dependencies = ["markdown_it_py", "lxml", "types-lxml", "pytest", "pytest-cov", "black", "flake8"]
requires-python = ">=3.10.7"

[tool.setuptools]
packages = ["content", "layout", "script", "static", "test"]
